sca_framework/requirements.txt
matplotlib==3.8.3
neptune==1.9.1
npy_append_array==0.9.16
numpy==1.26.4
omegaconf==2.3.0
picosdk==1.1
pyserial==3.5
pytorch_lightning==1.9.5
rescaler==0.0.3
scikit_learn==1.4.1.post1
scipy==1.12.0
torch==2.2.1
torchmetrics==1.3.1
tqdm==4.66.2
ipykernel==6.29.3
ipywidgets==8.1.0
